Convert to multi-project
So we can share code between the PostgreSQL and MySQL implementations.
