+++ postgresql/src/main/resources/reference.conf
@@ -1,0 +1,39 @@
+postgres-journal {
+  # FQCN of the R2DBC Postgres journal plugin
+  class = "akka.persistence.postgresql.journal.PostgresJournal"
+
+  db {
+    # The hostname of the database server
+    hostname = "localhost"
+
+    # The database username
+    username = "postgres"
+
+    # The username's password
+    password = "s3cr3t"
+
+    # The name of the database
+    database = "db"
+  }
+
+}
+
+postgres-snapshot {
+  # FQCN of the R2DBC Postgres snapshot store plugin
+  class = "akka.persistence.postgresql.snapshot.PostgresSnapshotStore"
+
+  db {
+    # The hostname of the database server
+    hostname = "localhost"
+
+    # The database username
+    username = "postgres"
+
+    # The username's password
+    password = "s3cr3t"
+
+    # The name of the database
+    database = "db"
+  }
+
+}
