+++ postgresql/src/main/scala/akka/persistence/postgresql/snapshot/SnapshotEntry.scala
@@ -1,0 +1,17 @@
+package akka.persistence.postgresql.snapshot
+
+import java.lang.{Long => JLong}
+
+object SnapshotEntry {
+
+  def of(pId: String, seqNr: JLong, timestamp: JLong, snapshot: Array[Byte]): SnapshotEntry =
+    SnapshotEntry(pId, seqNr.toLong, timestamp.toLong, snapshot)
+
+}
+
+final case class SnapshotEntry(
+    persistenceId: String,
+    sequenceNumber: Long,
+    timestamp: Long,
+    snapshot: Array[Byte]
+)
