--- src/main/scala/akka/persistence/r2dbc/snapshot/SnapshotEntry.scala
@@ -1,17 +1,0 @@
-package akka.persistence.r2dbc.snapshot
-
-import java.lang.{Long => JLong}
-
-object SnapshotEntry {
-
-  def of(pId: String, seqNr: JLong, timestamp: JLong, snapshot: Array[Byte]): SnapshotEntry =
-    SnapshotEntry(pId, seqNr.toLong, timestamp.toLong, snapshot)
-
-}
-
-final case class SnapshotEntry(
-    persistenceId: String,
-    sequenceNumber: Long,
-    timestamp: Long,
-    snapshot: Array[Byte]
-)
